refactor: Use general `parser.common` import for StrRegion and `.common`
