+++ parser/common/str_region.py
@@ -2,6 +2,8 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
+__all__ = ['StrRegion']
+
 
 @dataclass
 class StrRegion:
