+++ parser/cst/named_node.py
@@ -4,10 +4,10 @@ from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, overload, Sequence, cast, Literal
 
 from .base_node import Leaf, AnyNode, Node
-from parser.common.str_region import StrRegion
+from ..common import StrRegion
 
 if TYPE_CHECKING:
-    from parser.tokens import Token
+    from ..tokens import Token
 
 
 # NOTE: these classes only follow the Liskov substitution principle on
