+++ test/common.py
@@ -5,10 +5,10 @@ from __future__ import annotations
 from enum import IntFlag, Enum
 from typing import Sequence, TypeVar
 
-from parser.cst.base_node import Leaf, AnyNode, Node
-from parser.common.tree_print import tformat
-from parser.cst.treegen import TreeGen, CstParseError
 from parser.common.error import BaseParseError
+from parser.common.tree_print import tformat
+from parser.cst.base_node import Leaf, AnyNode, Node
+from parser.cst.treegen import TreeGen, CstParseError
 from parser.lexer import Tokenizer
 from parser.lexer.tokens import Token, OpToken
 from test.snapshottest import SnapshotTestCase
